Removed some compiler warnings, fix for OSX file chooser modality.

+++ modules/juce_gui_basics/misc/juce_BubbleComponent.h
@@ -144,7 +144,7 @@ public:
 
 private:
     Rectangle<int> content;
-    Point<float> arrowTip;
+    Point<int> arrowTip;
     int allowablePlacements;
     DropShadowEffect shadow;
